refactor: MainContent 컴포넌트를 AppRouter로 이름 변경
