+++ src/features/chat/index.js
@@ -1,0 +1,1 @@
+export { default as ChatPage } from "./pages/ChatPage";
