+++ src/features/local-companion/index.js
@@ -1,0 +1,3 @@
+export { default as LocalCompanionPage } from "./pages/LocalCompanionPage";
+
+export { default as CreateLocalModal } from "./modals/LocalCompanionPostModal";
