+++ src/features/planned-companion/index.js
@@ -1,0 +1,4 @@
+export { default as PlannedCompanionPage } from "./pages/PlannedCompanionPage";
+
+export { default as PlannedCompanionPostModal } from "./modals/PlannedCompanionPostModal";
+export { default as CreatePlannedModal } from "./modals/PlannedCompanionPostModal";
