+++ src/shared/index.js
@@ -7,6 +7,7 @@ export { default as Sidebar } from "./components/layout/Sidebar";
 export { default as Filterbar } from "./components/Filterbar";
 export { default as FallbackImage } from "./components/FallbackImage";
 export { default as ProfileImage } from "./components/ProfileImage";
+export { default as PostForm } from "./components/forms/PostForm";
 
 // ✅ Hooks
 export { default as useLockBodyScroll } from "./hooks/useLockBodyScroll";
@@ -16,3 +17,6 @@ export * from "./utils/regionMapper";
 
 // ✅ Constants
 export { default as regionOptions } from "./constants/regionOptions";
+
+// ✅ Pages
+export { default as ErrorPage } from "./pages/ErrorPage";
